Add behavior tree rendering to simulator visualizer
