+++ tools/twix/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod behavior_tree;
 pub mod twix_painter;
 pub mod zoom_and_pan;
